requirements.txt
numpy>=1.26
scipy>=1.12
matplotlib>=3.8
soundfile>=0.12
datasets[audio]>=2.18,<4.0
g2p-en>=2.1
PyYAML>=6.0
scikit-learn>=1.4
torch>=2.2
transformers>=4.40
sentencepiece>=0.2
